kxstudios_email_backend/requirements.txt

fastapi
uvicorn
python-dotenv
email-validator
python-multipart
